pyproject.toml
[project]
name = "demand-forecast-data-platform"
version = "0.1.0"
description = "A local-first data engineering platform for demand forecasting and inventory intelligence."
requires-python = "==3.11.*"
dependencies = [
    "psycopg[binary]",
    "python-dotenv",
    "pandas",
    "openpyxl",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests requiring a running PostgreSQL instance",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
